added check that desktop icons have been made before refreshing the desktop

+++ docker/launch.sh
@@ -721,11 +721,12 @@ else
     rm ${SAVE_LOGS_APP_FOLDER}/save_latest_ros_logs/customer.key   
 fi
 
-echo "Refreshing desktop..."
-if [[ $(cat /etc/*release | grep VERSION_CODENAME) = *"bionic"* ]]; then
-    #killall nautilus-desktop && nautilus-desktop &
-    killall nautilus-desktop &> /dev/null
-    nautilus-desktop &> /dev/null &
+if [[ $(cat /etc/*release | grep VERSION_CODENAME) = *"bionic"* && $(ps -ef | grep "nautilus-desktop" | grep -v "grep" | wc -l) ]]; then
+    if [[ ${DESKTOP_ICON} = true ||  ${DEMO_ICONS} = true ]]; then
+        echo "Refreshing desktop..."
+        killall nautilus-desktop &> /dev/null
+        nautilus-desktop &> /dev/null &
+    fi
 fi
 
 echo ""
